+++ react/IconPause.tsx
@@ -1,0 +1,9 @@
+import React from 'react'
+
+import Icon from './components/Icon'
+
+const IconPause = (props: IconProps) => {
+  return <Icon id="mpa-pause" handle="pauseIcon" {...props} />
+}
+
+export default IconPause