+++ react/IconVolumeOff.tsx
@@ -1,0 +1,9 @@
+import React from 'react'
+
+import Icon from './components/Icon'
+
+const IconVolumeOff = (props: IconProps) => {
+  return <Icon id="sti-volume-off" handle="volumeOffIcon" {...props} />
+}
+
+export default IconVolumeOff