+++ react/IconVolumeOn.tsx
@@ -1,0 +1,9 @@
+import React from 'react'
+
+import Icon from './components/Icon'
+
+const IconVolumeOn = (props: IconProps) => {
+  return <Icon id="sti-volume-on" handle="volumeOnIcon" {...props} />
+}
+
+export default IconVolumeOn